+++ packages/transports/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./otel.js";
 export * from "./http.js";
 export * from "./mongo.js";
+export * from "./loki.js";